# Reduce method calls in ExpressionTypeHolder