@staabm 2025-03-07
debugged the `src/PhpSpreadsheet/Reader/Xlsx.php` case further:

there is one union involved with more then 30 elements, which is resolved from the phpdoc `ZipArchive::ER_*` (from phpstan core signatures for ZipArchive).

the 2nd biggest union has 8 elements and is resolved from this phpdoc
https://github.com/PHPOffice/PhpSpreadsheet/blob/3ef8ec4b1a095f700412c6e44f15b3eaedae23fe/src/PhpSpreadsheet/Worksheet/Worksheet.php#L2713

-> there is no really huge union involved it seems. we spent most of the time with merging scope objects

profile can be seen here: https://blackfire.io/profiles/0293ba31-80fc-403f-9452-bfda15b7c7c7/graph

-> I am still investigating